fix prediction docs canonical route
