+++ app/docs/models/[model]/page.tsx
@@ -1,10 +1,11 @@
 import type { Metadata } from "next";
-import { notFound } from "next/navigation";
+import { notFound, redirect } from "next/navigation";
 
 import { DocsAddress } from "@/components/docs-address";
 import { DocsExternalLink } from "@/components/docs-external-link";
 import { DocsShell } from "@/components/docs-shell";
 import styles from "@/components/docs-experience.module.css";
+import { predictionMarketsDocsMetadata } from "@/components/prediction-markets-docs";
 
 type ModelSlug = "classic" | "custom" | "prediction-markets" | "stock-paired";
 
@@ -29,12 +30,6 @@ const customSections = [
   { id: "launch-information", label: "Launch information" },
   { id: "router-provenance", label: "Router provenance" },
   { id: "project-presentation", label: "What activation does not prove" },
-] as const;
-
-const predictionMarketSections = [
-  { id: "overview", label: "Overview" },
-  { id: "current-release", label: "Current release" },
-  { id: "boundaries", label: "Boundaries" },
 ] as const;
 
 const stockPairedSections = [
@@ -87,6 +82,7 @@ export async function generateMetadata({
 }): Promise<Metadata> {
   const { model } = await params;
   if (!isModelSlug(model)) return {};
+  if (model === "prediction-markets") return predictionMarketsDocsMetadata;
 
   const metadata = modelMetadata[model];
   return {
@@ -460,74 +456,6 @@ function CustomDocs() {
   );
 }
 
-function PredictionMarketsDocs() {
-  return (
-    <DocsShell
-      currentPath="/docs/models/prediction-markets"
-      kicker="Launch model"
-      title="Prediction Markets"
-      description="Programmable's open-source Uniswap v4 launch model for onchain outcome markets."
-      sections={predictionMarketSections}
-    >
-      <section id="overview">
-        <h2>Overview</h2>
-        <p>
-          Prediction Markets lets people create and trade onchain outcome
-          markets through Programmable. The product interface and the protocol
-          release are maintained separately so the current implementation can
-          evolve without turning this overview into a stale specification.
-        </p>
-        <div className={styles.sourceLinks}>
-          <DocsExternalLink
-            href="https://programmable.market/markets"
-            variant="chip"
-          >
-            Open Prediction Markets
-          </DocsExternalLink>
-          <DocsExternalLink
-            href="https://github.com/0xprogrammable/programmable-prediction-markets"
-            variant="chip"
-          >
-            Current source and release details
-          </DocsExternalLink>
-        </div>
-      </section>
-
-      <section id="current-release">
-        <h2>Use the current release</h2>
-        <p>
-          The canonical Prediction Markets repository is the source of truth for
-          the active protocol release. It defines:
-        </p>
-        <ul className={styles.contentList}>
-          <li>Current networks and supported market types.</li>
-          <li>Collateral and activation rules.</li>
-          <li>Fees, recipients and creator rewards.</li>
-          <li>Trading, resolution and payout rules.</li>
-          <li>Contract addresses, verified source and release evidence.</li>
-        </ul>
-        <div className={styles.callout}>
-          <strong>Do not copy release details from this overview.</strong>
-          <p>
-            Before creating, trading, integrating or describing a market, check
-            the current source and release record in the canonical repository.
-          </p>
-        </div>
-      </section>
-
-      <section id="boundaries">
-        <h2>Boundaries</h2>
-        <p>
-          A visible market, verified source or passing test does not guarantee
-          liquidity, execution at a chosen price, support in another application
-          or safety. Outcome assets can lose all value. Follow the current
-          security and integration guidance in the canonical repository.
-        </p>
-      </section>
-    </DocsShell>
-  );
-}
-
 function StockPairedDocs() {
   const contractRows = [
     {
@@ -825,9 +753,11 @@ export default async function ModelDocsPage({
 }) {
   const { model } = await params;
   if (!isModelSlug(model)) notFound();
+  if (model === "prediction-markets") {
+    redirect("/docs/tokens/prediction-markets");
+  }
 
   if (model === "classic") return <ClassicDocs />;
   if (model === "custom") return <CustomDocs />;
-  if (model === "prediction-markets") return <PredictionMarketsDocs />;
   return <StockPairedDocs />;
 }
