+++ app/docs/tokens/prediction-markets/page.tsx
@@ -1,0 +1,10 @@
+import {
+  PredictionMarketsDocs,
+  predictionMarketsDocsMetadata,
+} from "@/components/prediction-markets-docs";
+
+export const metadata = predictionMarketsDocsMetadata;
+
+export default function PredictionMarketsPage() {
+  return <PredictionMarketsDocs />;
+}
