+++ tests/prediction-market-docs.test.ts
@@ -10,13 +10,19 @@ const root = process.cwd();
 const read = (path: string) => readFileSync(join(root, path), "utf8");
 
 const modelPage = read("app/docs/models/[model]/page.tsx");
+const canonicalPage = read("app/docs/tokens/prediction-markets/page.tsx");
+const predictionDocsComponent = read("components/prediction-markets-docs.tsx");
 const gitBookPage = read("docs/public/models/prediction-markets.md");
 const readme = read("README.md");
 const styleGuide = read("docs/public/.gitbook/STYLEGUIDE.md");
-const predictionModelPage = modelPage.slice(
-  modelPage.indexOf("function PredictionMarketsDocs()"),
-  modelPage.indexOf("function StockPairedDocs()"),
-);
+const vercelConfig = JSON.parse(read("vercel.json")) as {
+  redirects?: Array<{
+    destination: string;
+    permanent: boolean;
+    source: string;
+  }>;
+};
+const predictionModelPage = predictionDocsComponent;
 const canonicalSource =
   "https://github.com/0xprogrammable/programmable-prediction-markets";
 const releaseBridgeSources = [
@@ -39,26 +45,39 @@ describe("Prediction Markets documentation", () => {
 
     expect(navigationEntries).toContainEqual(
       expect.objectContaining({
-        href: "/docs/models/prediction-markets",
+        href: "/docs/tokens/prediction-markets",
         label: "Prediction Markets",
       }),
     );
     expect(docsSearchItems).toContainEqual(
       expect.objectContaining({
-        href: "/docs/models/prediction-markets",
+        href: "/docs/tokens/prediction-markets",
         title: "Prediction Markets",
       }),
     );
     expect(sitemap().map((entry) => entry.url)).toContain(
-      "https://programmable.market/docs/models/prediction-markets",
+      "https://programmable.market/docs/tokens/prediction-markets",
     );
     expect(sitemap().map((entry) => entry.url)).toContain(
       "https://programmable.market/markets",
     );
+    expect(predictionDocsComponent).toContain(
+      'currentPath="/docs/tokens/prediction-markets"',
+    );
+    expect(canonicalPage).toContain("<PredictionMarketsDocs />");
     expect(modelPage).toContain(
-      'currentPath="/docs/models/prediction-markets"',
+      'if (model === "prediction-markets") return predictionMarketsDocsMetadata;',
+    );
+    expect(modelPage).toContain('redirect("/docs/tokens/prediction-markets")');
+    expect(predictionDocsComponent).toContain(
+      'canonical: "/docs/tokens/prediction-markets"',
     );
     expect(gitBookPage).toContain("# Prediction Markets");
+    expect(vercelConfig.redirects).toContainEqual({
+      destination: "/docs/tokens/prediction-markets",
+      permanent: true,
+      source: "/docs/models/prediction-markets",
+    });
   });
 
   it("keeps current release facts in the canonical source repository", () => {
